requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.15,<4.0
django-filter>=24.3,<25.0
drf-spectacular>=0.27,<1.0
python-dotenv>=1.0,<2.0